SnowflakeDestination is implementing the AbstractDestination class
