+++ airbyte-integrations/connectors/destination-snowflake/src/main/java/io/airbyte/integrations/destination/snowflake/SnowflakeDestination.java
@@ -25,15 +25,21 @@
 package io.airbyte.integrations.destination.snowflake;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.google.common.base.Charsets;
+import com.google.common.collect.Lists;
 import io.airbyte.commons.json.Jsons;
+import io.airbyte.commons.lang.CloseableQueue;
 import io.airbyte.commons.resources.MoreResources;
+import io.airbyte.integrations.base.AbstractDestination;
 import io.airbyte.integrations.base.Destination;
 import io.airbyte.integrations.base.DestinationConsumer;
 import io.airbyte.integrations.base.IntegrationRunner;
 import io.airbyte.integrations.base.SQLNamingResolvable;
+import io.airbyte.integrations.base.WriteConfig;
 import io.airbyte.protocol.models.AirbyteConnectionStatus;
 import io.airbyte.protocol.models.AirbyteConnectionStatus.Status;
 import io.airbyte.protocol.models.AirbyteMessage;
+import io.airbyte.protocol.models.AirbyteRecordMessage;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.ConfiguredAirbyteStream;
 import io.airbyte.protocol.models.ConnectorSpecification;
@@ -42,21 +48,28 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
 import java.util.function.Supplier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SnowflakeDestination implements Destination {
+public class SnowflakeDestination extends AbstractDestination implements Destination {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(SnowflakeDestination.class);
 
   protected static final String COLUMN_NAME = "data";
 
   private final SQLNamingResolvable namingResolver;
+  private Supplier<Connection> connectionFactory = null;
 
   public SnowflakeDestination() {
     namingResolver = new SnowflakeSQLNaming();
@@ -84,70 +97,114 @@ public class SnowflakeDestination implements Destination {
     return namingResolver;
   }
 
-  /**
-   * Strategy:
-   * <p>
-   * 1. Create a temporary table for each stream
-   * </p>
-   * <p>
-   * 2. Accumulate records in a buffer. One buffer per stream.
-   * </p>
-   * <p>
-   * 3. As records accumulate write them in batch to the database. We set a minimum numbers of records
-   * before writing to avoid wasteful record-wise writes.
-   * </p>
-   * <p>
-   * 4. Once all records have been written to buffer, flush the buffer and write any remaining records
-   * to the database (regardless of how few are left).
-   * </p>
-   * <p>
-   * 5. In a single transaction, delete the target tables if they exist and rename the temp tables to
-   * the final table name.
-   * </p>
-   *
-   * @param config - Snowflake-specific configuration object as json.
-   * @param catalog - schema of the incoming messages.
-   * @return consumer that writes singer messages to the database.
-   * @throws Exception - anything could happen!
-   */
   @Override
-  public DestinationConsumer<AirbyteMessage> write(JsonNode config, ConfiguredAirbyteCatalog catalog) throws Exception {
-    // connect to snowflake
-    final Supplier<Connection> connectionFactory = SnowflakeDatabase.getConnectionFactory(config);
-    Map<String, SnowflakeWriteContext> writeBuffers = new HashMap<>();
-    Set<String> schemaSet = new HashSet<>();
+  protected void connectDatabase(JsonNode config) {
+    connectionFactory = SnowflakeDatabase.getConnectionFactory(config);
+  }
 
-    // create temporary tables if they do not exist
-    // we don't use temporary/transient since we want to control the lifecycle
-    for (final ConfiguredAirbyteStream stream : catalog.getStreams()) {
-      final String streamName = stream.getStream().getName();
-      final String schemaName = getNamingResolver().getIdentifier(config.get("schema").asText());
-      final String tableName = getNamingResolver().getRawTableName(streamName);
-      final String tmpTableName = getNamingResolver().getTmpTableName(streamName);
-      if (!schemaSet.contains(schemaName)) {
-        final String query = String.format("CREATE SCHEMA IF NOT EXISTS %s;", schemaName);
+  @Override
+  protected void queryDatabase(String query) throws Exception {
+    SnowflakeDatabase.executeSync(connectionFactory, query);
+  }
 
-        SnowflakeDatabase.executeSync(connectionFactory, query);
-        schemaSet.add(schemaName);
-      }
-      final String query = String.format(
-          "CREATE TABLE IF NOT EXISTS %s.%s ( \n"
-              + "ab_id VARCHAR PRIMARY KEY,\n"
-              + "\"%s\" VARIANT,\n"
-              + "emitted_at TIMESTAMP WITH TIME ZONE DEFAULT current_timestamp()\n"
-              + ") data_retention_time_in_days = 0;",
-          schemaName, tmpTableName, COLUMN_NAME);
+  @Override
+  protected String createRawTableQuery(String schemaName, String streamName) {
+    return String.format(
+        "CREATE TABLE IF NOT EXISTS %s.%s ( \n"
+            + "ab_id VARCHAR PRIMARY KEY,\n"
+            + "\"%s\" VARIANT,\n"
+            + "emitted_at TIMESTAMP WITH TIME ZONE DEFAULT current_timestamp()\n"
+            + ") data_retention_time_in_days = 0;",
+        schemaName, streamName, COLUMN_NAME);
+  }
 
-      SnowflakeDatabase.executeSync(connectionFactory, query);
+  @Override
+  public void writeQuery(int batchSize, CloseableQueue<byte[]> writeBuffer, String schemaName, String tmpTableName) {
+    final List<AirbyteRecordMessage> records = accumulateRecordsFromBuffer(writeBuffer, batchSize);
 
-      final Path queueRoot = Files.createTempDirectory("queues");
-      final BigQueue writeBuffer = new BigQueue(queueRoot.resolve(stream.getStream().getName()), stream.getStream().getName());
-      writeBuffers.put(stream.getStream().getName(), new SnowflakeWriteContext(schemaName, tableName, tmpTableName, writeBuffer));
+    LOGGER.info("max size of batch: {}", batchSize);
+    LOGGER.info("actual size of batch: {}", records.size());
+
+    if (records.isEmpty()) {
+      return;
     }
 
-    // write to transient tables
-    // if success copy delete main table if exists. rename transient tables to real tables.
-    return new SnowflakeRecordConsumer(connectionFactory, writeBuffers, catalog);
+    try (final Connection conn = connectionFactory.get()) {
+      // Strategy: We want to use PreparedStatement because it handles binding values to the SQL query
+      // (e.g. handling formatting timestamps). A PreparedStatement statement is created by supplying the
+      // full SQL string at creation time. Then subsequently specifying which values are bound to the
+      // string. Thus there will be two loops below.
+      // 1) Loop over records to build the full string.
+      // 2) Loop over the records and bind the appropriate values to the string.
+      final StringBuilder sql = new StringBuilder().append(String.format(
+          "INSERT INTO %s.%s (ab_id, \"%s\", emitted_at) SELECT column1, parse_json(column2), column3 FROM VALUES\n",
+          schemaName,
+          tmpTableName,
+          SnowflakeDestination.COLUMN_NAME));
+
+      // first loop: build SQL string.
+      records.forEach(r -> sql.append("(?, ?, ?),\n"));
+      final String s = sql.toString();
+      final String s1 = s.substring(0, s.length() - 2) + ";";
+
+      try (final PreparedStatement statement = conn.prepareStatement(s1)) {
+        // second loop: bind values to the SQL string.
+        int i = 1;
+        for (final AirbyteRecordMessage message : records) {
+          // 1-indexed
+          statement.setString(i, UUID.randomUUID().toString());
+          statement.setString(i + 1, Jsons.serialize(message.getData()));
+          statement.setTimestamp(i + 2, Timestamp.from(Instant.ofEpochMilli(message.getEmittedAt())));
+          i += 3;
+        }
+
+        statement.execute();
+      }
+    } catch (SQLException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  private static List<AirbyteRecordMessage> accumulateRecordsFromBuffer(CloseableQueue<byte[]> writeBuffer, int maxRecords) {
+    final List<AirbyteRecordMessage> records = Lists.newArrayList();
+    for (int i = 0; i < maxRecords; i++) {
+      final byte[] record = writeBuffer.poll();
+      if (record == null) {
+        break;
+      }
+      final AirbyteRecordMessage message = Jsons.deserialize(new String(record, Charsets.UTF_8), AirbyteRecordMessage.class);
+      records.add(message);
+    }
+
+    return records;
+  }
+
+  @Override
+  public void commitRawTables(Map<String, WriteConfig> writeConfigs) throws Exception {
+    final StringBuilder query = new StringBuilder();
+    query.append("BEGIN;");
+    for (final WriteConfig writeContext : writeConfigs.values()) {
+      query.append(String.format("DROP TABLE IF EXISTS %s.%s;\n", writeContext.getSchemaName(), writeContext.getTableName()));
+      query.append(String.format("ALTER TABLE %s.%s RENAME TO %s.%s;\n", writeContext.getSchemaName(), writeContext.getTmpTableName(),
+          writeContext.getSchemaName(), writeContext.getTableName()));
+    }
+    query.append("COMMIT;");
+
+    final String renameQuery = query.toString();
+
+    SnowflakeDatabase.executeSync(connectionFactory, renameQuery, true, rs -> null);
+  }
+
+  @Override
+  public void cleanupTmpTables(Map<String, WriteConfig> writeConfigs) {
+    for (WriteConfig writeContext : writeConfigs.values()) {
+      try {
+        SnowflakeDatabase.executeSync(connectionFactory,
+            String.format("DROP TABLE IF EXISTS %s.%s;", writeContext.getSchemaName(), writeContext.getTmpTableName()));
+      } catch (SQLException | InterruptedException e) {
+        throw new RuntimeException(e);
+      }
+    }
   }
 
   public static void main(String[] args) throws Exception {
